% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_query_database.R
\name{anonymize_uuid}
\alias{anonymize_uuid}
\title{Anonymize user_id field}
\usage{
anonymize_uuid(.data)
}
\arguments{
\item{.data}{a data.frame object}
}
\value{
.data with anonymized user_id
}
\description{
Anonymize user_id field
}
