% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R
\name{app_server}
\alias{app_server}
\title{The application server-side}
\usage{
app_server(input, output, session)
}
\arguments{
\item{input, output, session}{Internal parameters for {shiny}.
DO NOT REMOVE.}

\item{row}{the row where you made the change
Update a row}

\item{data}{the data including your updated row
Delete a row}
}
\description{
The application server-side
}
