% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_connect_database.R
\docType{data}
\name{COLLECTIONS}
\alias{COLLECTIONS}
\title{Collections in an e-mission MongoDB database}
\format{
An object of class \code{character} of length 12.
}
\usage{
COLLECTIONS
}
\description{
There are 7 collections that exist in a e-mission data.

* Stage_Profiles
* Stage_analysis_timeseries
* Stage_pipeline_state
* Stage_timeseries
* Stage_timeseries_error
* Stage_usercache
* Stage_uuids
}
\keyword{datasets}
