% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{combine_named_lists}
\alias{combine_named_lists}
\title{Combine two named lists}
\usage{
combine_named_lists(x, y, replace = TRUE)
}
\arguments{
\item{x, }{y a named list. All the elements must be uniquely named.}

\item{replace}{a logical value, default as TRUE. If TRUE, all
the `x` elements with names that exist in the `y` list will be
replaced with the values in `y`.}
}
\description{
Combine two named lists
}
