% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_connect_database.R
\name{connect_stage_collections}
\alias{connect_stage_collections}
\title{Connect to the e-mission's MongoDB database}
\usage{
connect_stage_collections(
  db = "Stage_database",
  collections = COLLECTIONS,
  url = getOption("emdash.mongo_url")
)
}
\arguments{
\item{db}{database name. Default as 'Stage_database'.}

\item{collections}{collections to connect to, by default this try to connect to
all the collections in [COLLECTIONS].}

\item{url}{address of the emisison mongodb server in mongo
connection string [URI format](https://docs.mongodb.com/manual/reference/connection-string/).
The default is `getOption('emdash.mongo_url')`}
}
\value{
a list of mongolite connection objects.
}
\description{
Connect to the e-mission's MongoDB database
}
