% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_tidy_data.R
\name{convert_columns_to_datetime}
\alias{convert_columns_to_datetime}
\title{Convert columns to datetime}
\usage{
convert_columns_to_datetime(.data, cols, tz = "Australia/Sydney")
}
\arguments{
\item{.data}{a data.table object}

\item{cols}{the names of the columns to convert to datetime}

\item{tz}{time zone default as "Australia/Sydney".}
}
\value{
.data
}
\description{
Convert columns to datetime
}
