% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_query_database.R
\name{convert_datetime_string_to_datetime}
\alias{convert_datetime_string_to_datetime}
\title{Convert character columns to datetime columns}
\usage{
convert_datetime_string_to_datetime(.data, cols, tz = "Australia/Sydney")
}
\arguments{
\item{.data}{a data.frame.}

\item{cols}{columns to convert to datetime columns.}

\item{tz}{time zone. Default as "Australia/Sydney".}
}
\value{
.data
}
\description{
Convert character columns to datetime columns
}
