% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_update_insert_delete.R
\name{db_insert}
\alias{db_insert}
\title{Inserts a row as a document into the specified collection}
\usage{
db_insert(cons, collection_name, df_row)
}
\arguments{
\item{cons}{connection to mongodb}

\item{collection_name}{the collection to look in}

\item{df_row}{the dataframe row to insert}
}
\description{
Inserts a row as a document into the specified collection
}
