% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_update_insert_delete.R
\name{db_update}
\alias{db_update}
\title{Updates a document with the row 'df_row'}
\usage{
db_update(cons, collection_name, df_row)
}
\arguments{
\item{cons}{connection to mongodb}

\item{collection_name}{the collection to look in}

\item{df_row}{the dataframe row to update with}
}
\description{
Updates a document with the row 'df_row'
}
