% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_tidy_data.R
\name{generate_trajectories}
\alias{generate_trajectories}
\title{Tidy together the 'tidied trips' and 'tidied locations' to generate trajectories within a trip}
\usage{
generate_trajectories(tidied_trips, tidied_locations, project_crs = 4326)
}
\arguments{
\item{tidied_trips}{a data.frame from `tidy_cleaned_trips()`.}

\item{tidied_locations}{a tibble from `tidy_cleaned_locations()`.}

\item{project_crs}{a EPSG code. Default as 4326.}
}
\value{
a data.frame of class sf.
}
\description{
Tidy together the 'tidied trips' and 'tidied locations' to generate trajectories within a trip
}
