% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_DTedit.R
\name{mod_DTedit_server}
\alias{mod_DTedit_server}
\title{DTedit Server Function}
\usage{
mod_DTedit_server(
  input,
  output,
  session,
  table_data,
  Table_Type,
  suppl_table_sublist,
  DT_options,
  cons
)
}
\arguments{
\item{tab_name}{table name
Insert a row. "Create"}

\item{row}{the row where you made the change
Update a row}

\item{data}{the data including your updated row
Delete a row}
}
\description{
DTedit server module.
}
