% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_mapview.R
\name{mod_mapview_server}
\alias{mod_mapview_server}
\title{mapview Server Function}
\usage{
mod_mapview_server(
  input,
  output,
  session,
  data_sf,
  mapview.map.types = "CartoDB.DarkMatter"
)
}
\arguments{
\item{data_sf}{a spatial data object from the 'sf' package.}

\item{mapview.map.types}{see `map.types` in the [mapview::mapview] function.}
}
\description{
mapview Server Function
}
