% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_mapview.R
\name{mod_mapview_ui}
\alias{mod_mapview_ui}
\title{mapview UI Function}
\usage{
mod_mapview_ui(id, height)
}
\arguments{
\item{id}{Internal parameters for {shiny}.}

\item{height}{[leaflet::leafletOutput]'s `height` argument.}
}
\description{
A shiny Module.
}
