% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_query_database.R
\name{normalise_uuid}
\alias{normalise_uuid}
\title{Normalise UUID}
\usage{
normalise_uuid(.data, keep_uuid = FALSE)
}
\arguments{
\item{.data}{a data.frame created using `query_*` functions in `R/utils_query_database.R`.}

\item{keep_uuid}{a logical value default as FALSE. If this is true then the
original `uuid` field will not be removed from `.data`.}
}
\value{
a data.frame
}
\description{
Note that this function will rename `uuid` to `user_id`.
}
