% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_query_database.R
\name{query}
\alias{query}
\alias{query_cleaned_locations}
\alias{query_cleaned_locations_by_timestamp}
\alias{query_server_calls}
\alias{query_usercache_get_summ}
\alias{query_usercache_put_summ}
\alias{query_diary_summ}
\alias{query_cleaned_trips}
\alias{query_cleaned_trips_by_timestamp}
\alias{get_query_size}
\alias{query_cleaned_place}
\alias{query_cleaned_section}
\alias{query_raw_trips}
\alias{query_stage_uuids}
\alias{query_stage_profiles}
\title{Query functions}
\usage{
query_cleaned_locations(cons)

query_cleaned_locations_by_timestamp(cons, dates)

query_server_calls(cons)

query_usercache_get_summ(cons)

query_usercache_put_summ(cons)

query_diary_summ(cons)

query_cleaned_trips(cons)

query_cleaned_trips_by_timestamp(cons, dates)

get_query_size(cons, dates)

query_cleaned_place(cons)

query_cleaned_section(cons)

query_raw_trips(cons)

query_stage_uuids(cons)

query_stage_profiles(cons)
}
\arguments{
\item{cons}{a list of connections created with [connect_stage_collections()].}
}
\value{
a data.frame/data.table.
}
\description{
queries trips data between the start timestamp of the first date 
             and the start timestamp of the second date.

Returns the number of cleaned trip documents in between two dates
}
