% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_query_database.R
\name{query_manual_survey_response}
\alias{query_manual_survey_response}
\alias{query_time_use_response}
\alias{query_trip_confirm_response}
\title{Get manual survey responses}
\usage{
query_manual_survey_response(cons)

query_time_use_response(cons)

query_trip_confirm_response(cons)
}
\value{
a data.table object
}
\description{
This function returns the manual survey response objects stored in
the 'Stage_timeseries' collection. These are Enketo responses that
emTripLog and its Forks applications stored.
}
