% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_query_database.R
\name{query_max_trip_timestamp}
\alias{query_max_trip_timestamp}
\title{Finds the maximum trip end timestamp.}
\usage{
query_max_trip_timestamp(cons)
}
\description{
Finds the maximum trip end timestamp.
}
