% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_query_database.R
\name{query_min_trip_timestamp}
\alias{query_min_trip_timestamp}
\title{Finds the minimum trip start timestamp.}
\usage{
query_min_trip_timestamp(cons)
}
\description{
Finds the minimum trip start timestamp.
}
