% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{save_config_file}
\alias{save_config_file}
\title{Save the default app config file}
\usage{
save_config_file(dir_out)
}
\arguments{
\item{dir_out}{the directory to save the config file}
}
\value{
a logical value, `TRUE` if the file has been successfully saved.
}
\description{
This function saves the default app configuration file required by [run_app()]. 
Alternatively, you can download the file from 
https://github.com/asiripanich/emdash/blob/master/inst/config-default.yml.
}
