% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_tidy_data.R
\name{summarise_server_calls}
\alias{summarise_server_calls}
\title{Create a summary of server calls in data.table format.}
\usage{
summarise_server_calls(participants, server_calls)
}
\arguments{
\item{participants}{the output from `tidy_participants()`.}

\item{trips}{the output from `tidy_cleaned_trips()`.}
}
\value{
a data.table.
}
\description{
Create a summary of server calls in data.table format.
}
