% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_tidy_data.R
\name{summarise_trips}
\alias{summarise_trips}
\title{Create a summary of trips in data.table format.}
\usage{
summarise_trips(participants, trips)
}
\arguments{
\item{participants}{the output from `tidy_participants()`.}

\item{trips}{the output from `tidy_cleaned_trips()`.}
}
\value{
a data.table.
}
\description{
Create a summary of trips in data.table format.
}
