% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_tidy_data.R
\name{summarise_trips_without_trips}
\alias{summarise_trips_without_trips}
\title{summarise trip patterns of the participants}
\usage{
summarise_trips_without_trips(participants, cons)
}
\description{
summarise trip patterns of the participants
}
