% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_tidy_data.R
\name{tidy_cleaned_locations}
\alias{tidy_cleaned_locations}
\title{Tidy the 'cleaned locations' data.frame into a tibble.}
\usage{
tidy_cleaned_locations(cleaned_locations)
}
\arguments{
\item{cleaned_locations}{a data.table output from `query_cleaned_trips()`.}
}
\value{
a tibble.
}
\description{
Tidy the 'cleaned locations' data.frame into a tibble.
}
