% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_tidy_data.R
\name{tidy_cleaned_trips}
\alias{tidy_cleaned_trips}
\title{Tidy the 'cleaned trips' data.frame into a sf object.}
\usage{
tidy_cleaned_trips(
  cleaned_trips,
  project_crs = 4326,
  smallest_rounding_digit = 2
)
}
\arguments{
\item{cleaned_trips}{a data.table output from `query_cleaned_trips()`.}

\item{project_crs}{a EPSG code. Default as 4326.}

\item{smallest_rounding_digit}{an integer value.}
}
\value{
a spatial data.frame of class sf.
}
\description{
Tidy the 'cleaned trips' data.frame into a sf object.
}
