% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_tidy_data.R
\name{tidy_participants}
\alias{tidy_participants}
\title{Create a tidied participant data.table}
\usage{
tidy_participants(stage_profiles, stage_uuids)
}
\arguments{
\item{stage_profiles}{the output of `query_stage_profiles()`.}

\item{stage_uuids}{the output of `query_stage_uuids()`.}
}
\value{
a data.table
}
\description{
Create a tidied participant data.table
}
\note{
the `update_ts` field can be used to infer the signup datetime of each user.
}
