% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_tidy_data.R
\name{tidy_server_calls}
\alias{tidy_server_calls}
\title{Tidy the 'server calls' data.frame into a tibble.}
\usage{
tidy_server_calls(server_calls)
}
\arguments{
\item{server_calls}{a data.table output from `query_cleaned_trips()`.}
}
\value{
a tibble.
}
\description{
Tidy the 'server calls' data.frame into a tibble.
}
