% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_update_insert_delete.R
\name{db_delete}
\alias{db_delete}
\title{Deletes a document from Checkinout corresponding to df_row}
\usage{
db_delete(cons, collection_name, df_row)
}
\description{
Deletes a document from Checkinout corresponding to df_row
}
