% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_update_insert_delete.R
\name{db_insert}
\alias{db_insert}
\title{Inserts a row as a document into the specified collection}
\usage{
db_insert(cons, collection_name, df)
}
\description{
Inserts a row as a document into the specified collection
}
