% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_DTedit.R
\name{mod_DTedit_server}
\alias{mod_DTedit_server}
\title{DTedit Server Function}
\usage{
mod_DTedit_server(
  input,
  output,
  session,
  table_data,
  Table_Type,
  suppl_table_sublist,
  DT_options,
  cons
)
}
\arguments{
\item{input, output, session}{Internal parameters for {shiny}}

\item{table_data}{data to place in the editable data table}

\item{Table_Type}{Name of the mongo collection the data table is associated with}

\item{suppl_table_sublist}{list containing information related to the table}

\item{DT_options}{options used by DT::renderDataTable within dtedit}

\item{cons}{the connection to mongodb}

\item{data}{a data.frame}
}
\description{
DTedit Server Function
}
