% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_DTedit.R
\name{mod_DTedit_server}
\alias{mod_DTedit_server}
\title{DTedit Server Function}
\usage{
mod_DTedit_server(input, output, session, data)
}
\arguments{
\item{data}{a data.frame}

\item{tab_name}{table name}
}
\description{
DTedit Server Function
}
