% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_tidy_data.R
\name{normalise_uuid}
\alias{normalise_uuid}
\title{Normalise UUID}
\usage{
normalise_uuid(.data, keep_uuid = FALSE)
}
\arguments{
\item{.data}{a data.frame.}

\item{keep_uuid}{logical.}
}
\value{
a data.frame
}
\description{
Normalise UUID
}
