% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_query_database.R
\name{query}
\alias{query}
\alias{query_cleaned_locations}
\alias{query_cleaned_locations_by_timestamp}
\alias{downsample_cleaned_locations_by_timestamp}
\alias{query_server_calls}
\alias{query_usercache_get_summ}
\alias{query_usercache_put_summ}
\alias{query_diary_summ}
\alias{query_cleaned_trips}
\alias{query_most_recent_n_trip_docs}
\alias{query_confirmed_trips_by_timestamp}
\alias{query_max_trip_timestamp}
\alias{query_min_trip_timestamp}
\alias{get_n_trips_in_query}
\alias{get_n_locations_in_query}
\alias{query_cleaned_place}
\alias{query_cleaned_section}
\alias{query_raw_trips}
\alias{query_stage_uuids}
\alias{query_stage_profiles}
\title{Query functions}
\usage{
query_cleaned_locations(cons)

query_cleaned_locations_by_timestamp(cons, dates)

downsample_cleaned_locations_by_timestamp(cons, dates, sample_size)

query_server_calls(cons)

query_usercache_get_summ(cons)

query_usercache_put_summ(cons)

query_diary_summ(cons)

query_usercache_get_summ(cons)

query_usercache_put_summ(cons)

query_diary_summ(cons)

query_cleaned_trips(cons)

query_most_recent_n_trip_docs(cons, n)

query_confirmed_trips_by_timestamp(cons, dates)

query_max_trip_timestamp(cons)

query_min_trip_timestamp(cons)

get_n_trips_in_query(cons, dates)

get_n_locations_in_query(cons, dates)

query_cleaned_place(cons)

query_cleaned_section(cons)

query_raw_trips(cons)

query_stage_uuids(cons, user_emails = NULL)

query_stage_profiles(cons, user_id)
}
\arguments{
\item{cons}{a list of connections created with [connect_stage_collections()].}

\item{dates}{a date range, a character vector of length two.
Dates must be in this format "yyyy-mm-dd".}

\item{sample_size}{number of location documents to be queried.}

\item{n}{number of trips used by `query_most_recent_n_trip_docs()`}

\item{user_emails}{user_email values in the Stage UUID collection to be filtered.}

\item{uuid}{a character vector of user IDs to be queried.}
}
\value{
a data.frame/data.table.

`downsample_cleaned_locations_by_timestamp` returns a dataframe containing
`sample_size` location documents within the date range.

`query_usercache_put_summ()` returns the first and last
 put calls for each user.

`query_most_recent_n_trip_docs()` yields trips data for the most recent n
trips, based on data.end_ts

`query_confirmed_trips_by_timestamp()` trips data
 between the start timestamp of the first date
 and the start timestamp of the second date.

`get_n_trips_in_query()` returns the number of cleaned
 trip documents in between two dates
}
\description{
Returns the number of cleaned locations documents in between two dates
}
