% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_query_database.R
\name{query}
\alias{query}
\alias{query_cleaned_locations}
\alias{query_server_calls}
\alias{query_cleaned_trips}
\alias{query_checkinout}
\alias{query_supplementary}
\alias{query_cleaned_place}
\alias{query_cleaned_section}
\alias{query_raw_trips}
\alias{query_stage_uuids}
\alias{query_stage_profiles}
\title{Query functions}
\usage{
query_cleaned_locations(cons)

query_server_calls(cons)

query_cleaned_trips(cons)

query_checkinout(cons)

query_supplementary(cons, name)

query_cleaned_place(cons)

query_cleaned_section(cons)

query_raw_trips(cons)

query_stage_uuids(cons)

query_stage_profiles(cons)
}
\arguments{
\item{cons}{a list of connections created with [connect_stage_collections()].}
}
\value{
a data.frame/data.table.
}
\description{
Query functions
}
